+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:theme="@style/Theme.YaimafuniAndroid"
         >
         <activity
-            android:name=".MainActivity"
+            android:name="com.ikemura23.yaeyama.linerchecker.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.YaimafuniAndroid"
             >
